# [WIP]  core: add global-scoped dispatcher 