--- a/tracing-core/src/dispatch.rs
+++ b/tracing-core/src/dispatch.rs
@@ -372,6 +372,9 @@ pub fn set_global_scoped(dispatcher: Dispatch) -> DefaultGuard {
     } else {
         None
     };
+
+    EXISTS.store(true, Ordering::Release);
+
     DefaultGuard(GuardInner::Global { prev, state })
 }