This branch adds the ability to set a scoped dispatcher for all threads
globally, which is unset when a RAII guard is dropped.

The global scoped default is stored in an `RwLock`. However, the
overhead of accessing the `RwLock` should be avoided when the global
scoped dispatcher has never been set. Therefore, this new API is *not* a
replacement for `set_global_default`, as accessing the scoped global
default dispatcher will have increased performance overhead versus
accessing the static global default. Additionally, the scoped default
dispatcher may only be used when the `std` feature is enabled, while the
unscoped static global default may be used in `no_std` environments.

Thus, the new `set_global_scoped` API should only be used when it's
necessary to set and unset multiple global dispatchers, and
`set_global_default` should be used in all other cases.

This is a draft because:
- I have yet to write documentation for the new APIs
- It's unclear whether the ability to do this is really worth the
  complexity added in `tracing-core`, both in terms of the amount of
  code in the `dispatch` module and in terms of API complexity (there
  are now 4 different functions for setting a dispatcher...)

<details>
<summary>criterion benchmark results</summary>

## Before

```
     Running unittests (target/release/deps/global_subscriber-644d3e647b65d97a)
WARNING: HTML report generation will become a non-default optional feature in Criterion.rs 0.4.0.
This feature is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'html_reports' feature in your Cargo.toml.

WARNING: In Criterion.rs 0.4.0, running criterion benchmarks outside of cargo-criterion will become a default optional feature.
The statistical analysis and reporting is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'cargo_bench_support' feature in your Cargo.toml.

Gnuplot not found, using plotters backend
global/collector/span_no_fields
                        time:   [13.806 ns 13.843 ns 13.882 ns]
Found 29 outliers among 100 measurements (29.00%)
  8 (8.00%) low severe
  5 (5.00%) low mild
  7 (7.00%) high mild
  9 (9.00%) high severe
global/collector/event  time:   [18.363 ns 18.425 ns 18.479 ns]
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
global/collector/enter_span
                        time:   [2.8698 ns 2.8747 ns 2.8818 ns]
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) high mild
  5 (5.00%) high severe
global/collector/span_repeatedly
                        time:   [3.4762 us 3.4833 us 3.4905 us]
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high severe
global/collector/span_with_fields
                        time:   [46.785 ns 46.856 ns 46.927 ns]
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low severe
  3 (3.00%) low mild
  1 (1.00%) high severe

global/dispatch/get_ref time:   [2.0163 ns 2.0191 ns 2.0226 ns]
Found 9 outliers among 100 measurements (9.00%)
  2 (2.00%) low mild
  3 (3.00%) high mild
  4 (4.00%) high severe
global/dispatch/get_clone
                        time:   [8.9843 ns 8.9896 ns 8.9956 ns]
Found 13 outliers among 100 measurements (13.00%)
  4 (4.00%) low severe
  4 (4.00%) low mild
  2 (2.00%) high mild
  3 (3.00%) high severe

     Running unittests (target/release/deps/no_subscriber-614ea688dc9c8eba)
WARNING: HTML report generation will become a non-default optional feature in Criterion.rs 0.4.0.
This feature is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'html_reports' feature in your Cargo.toml.

WARNING: In Criterion.rs 0.4.0, running criterion benchmarks outside of cargo-criterion will become a default optional feature.
The statistical analysis and reporting is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'cargo_bench_support' feature in your Cargo.toml.

Gnuplot not found, using plotters backend
no_subscriber/span      time:   [1.8681 ns 1.8753 ns 1.8827 ns]
Found 11 outliers among 100 measurements (11.00%)
  10 (10.00%) high mild
  1 (1.00%) high severe
no_subscriber/event     time:   [237.01 ps 237.38 ps 237.82 ps]
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe
no_subscriber/relaxed_load
                        time:   [487.80 ps 492.57 ps 497.96 ps]
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
no_subscriber/acquire_load
                        time:   [462.68 ps 463.81 ps 465.03 ps]
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe
no_subscriber/log       time:   [236.10 ps 238.26 ps 240.70 ps]
Found 10 outliers among 100 measurements (10.00%)
  7 (7.00%) high mild
  3 (3.00%) high severe

no_subscriber_field/span
                        time:   [1.8620 ns 1.8661 ns 1.8709 ns]
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe
no_subscriber_field/event
                        time:   [289.92 ps 291.19 ps 293.10 ps]
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
no_subscriber_field/log time:   [466.41 ps 467.87 ps 469.66 ps]
Found 10 outliers among 100 measurements (10.00%)
  4 (4.00%) low mild
  3 (3.00%) high mild
  3 (3.00%) high severe

     Running unittests (target/release/deps/subscriber-51b0d2495c9ed443)
WARNING: HTML report generation will become a non-default optional feature in Criterion.rs 0.4.0.
This feature is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'html_reports' feature in your Cargo.toml.

WARNING: In Criterion.rs 0.4.0, running criterion benchmarks outside of cargo-criterion will become a default optional feature.
The statistical analysis and reporting is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'cargo_bench_support' feature in your Cargo.toml.

Gnuplot not found, using plotters backend
scoped/subscriber/span_no_fields
                        time:   [21.499 ns 21.640 ns 21.766 ns]
scoped/subscriber/enter_span
                        time:   [3.2577 ns 3.2678 ns 3.2791 ns]
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  1 (1.00%) high severe
scoped/subscriber/event time:   [16.280 ns 16.326 ns 16.378 ns]
Found 10 outliers among 100 measurements (10.00%)
  1 (1.00%) low mild
  6 (6.00%) high mild
  3 (3.00%) high severe
scoped/subscriber/span_repeatedly
                        time:   [3.7232 us 3.7301 us 3.7378 us]
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe
scoped/subscriber/span_with_fields
                        time:   [57.679 ns 57.723 ns 57.773 ns]
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  2 (2.00%) high severe
scoped/subscriber/span_with_fields_record
                        time:   [323.68 ns 323.97 ns 324.29 ns]
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe

no/dispatch/get_ref     time:   [2.3818 ns 2.3856 ns 2.3894 ns]
Found 14 outliers among 100 measurements (14.00%)
  2 (2.00%) low severe
  1 (1.00%) low mild
  2 (2.00%) high mild
  9 (9.00%) high severe
no/dispatch/get_clone   time:   [8.9051 ns 8.9341 ns 8.9659 ns]
Found 9 outliers among 100 measurements (9.00%)
  3 (3.00%) low severe
  3 (3.00%) low mild
  1 (1.00%) high mild
  2 (2.00%) high severe

scoped/dispatch/get_ref time:   [3.9818 ns 3.9948 ns 4.0098 ns]
Found 9 outliers among 100 measurements (9.00%)
  2 (2.00%) low mild
  1 (1.00%) high mild
  6 (6.00%) high severe
scoped/dispatch/get_clone
                        time:   [17.489 ns 17.514 ns 17.540 ns]

     Running unittests (target/release/deps/tracing_appender-4fcf629780e34446)

running 18 tests
test non_blocking::test::backpressure_exerted ... ignored
test non_blocking::test::logs_dropped_if_lossy ... ignored
test non_blocking::test::multi_threaded_writes ... ignored
test rolling::test::test_join_date_never ... ignored
test rolling::test::test_next_date_daily ... ignored
test rolling::test::test_next_date_hourly ... ignored
test rolling::test::test_next_date_minutely ... ignored
test rolling::test::test_next_date_never ... ignored
test rolling::test::test_rotation_path_daily ... ignored
test rolling::test::test_rotation_path_hourly ... ignored
test rolling::test::test_rotation_path_minutely ... ignored
test rolling::test::test_round_date_daily ... ignored
test rolling::test::test_round_date_hourly ... ignored
test rolling::test::test_round_date_minutely ... ignored
test rolling::test::write_daily_log ... ignored
test rolling::test::write_hourly_log ... ignored
test rolling::test::write_minutely_log ... ignored
test rolling::test::write_never_log ... ignored

test result: ok. 0 passed; 0 failed; 18 ignored; 0 measured; 0 filtered out; finished in 0.00s

     Running unittests (target/release/deps/tracing_attributes-1e31b4a0f226ab0d)

running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.00s

     Running unittests (target/release/deps/tracing_core-587538a453cdaedf)

running 21 tests
test callsite::tests::linked_list_empty ... ignored
test callsite::tests::linked_list_push ... ignored
test callsite::tests::linked_list_push_several ... ignored
test callsite::tests::linked_list_repeated - should panic ... ignored
test dispatch::test::default_dispatch ... ignored
test dispatch::test::default_no_collector ... ignored
test dispatch::test::dispatch_downcasts ... ignored
test dispatch::test::dispatch_is ... ignored
test dispatch::test::events_dont_infinite_loop ... ignored
test dispatch::test::spans_dont_infinite_loop ... ignored
test field::test::empty_fields_are_skipped ... ignored
test field::test::empty_value_set_is_empty ... ignored
test field::test::fields_from_other_callsets_are_skipped ... ignored
test field::test::record_debug_fn ... ignored
test field::test::sparse_value_sets_are_not_empty ... ignored
test field::test::value_set_with_no_values_is_empty ... ignored
test field::test::value_sets_with_fields_from_other_callsites_are_empty ... ignored
test metadata::tests::filter_level_conversion ... ignored
test metadata::tests::level_filter_is_usize_sized ... ignored
test metadata::tests::level_filter_reprs ... ignored
test metadata::tests::level_from_str ... ignored

test result: ok. 0 passed; 0 failed; 21 ignored; 0 measured; 0 filtered out; finished in 0.00s

     Running unittests (target/release/deps/tracing_error-528504973cad270e)

running 3 tests
test backtrace::tests::capture_empty ... ignored
test backtrace::tests::capture_supported ... ignored
test backtrace::tests::capture_unsupported ... ignored

test result: ok. 0 passed; 0 failed; 3 ignored; 0 measured; 0 filtered out; finished in 0.00s

     Running unittests (target/release/deps/tracing_flame-54ac419cdf5736fe)

running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.00s

     Running unittests (target/release/deps/tracing_futures-29e65b5f8c0d408f)

running 4 tests
test tests::futures_01_tests::future_enter_exit_is_reasonable ... ignored
test tests::futures_01_tests::future_error_ends_span ... ignored
test tests::futures_01_tests::span_follows_future_onto_threadpool ... ignored
test tests::futures_01_tests::stream_enter_exit_is_reasonable ... ignored

test result: ok. 0 passed; 0 failed; 4 ignored; 0 measured; 0 filtered out; finished in 0.00s

     Running unittests (target/release/deps/tracing_journald-375256083c8e6484)

running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.00s

     Running unittests (target/release/deps/tracing_log-3da796205da99c8b)

running 5 tests
test test::debug_callsite_is_correct ... ignored
test test::error_callsite_is_correct ... ignored
test test::info_callsite_is_correct ... ignored
test test::trace_callsite_is_correct ... ignored
test test::warn_callsite_is_correct ... ignored

test result: ok. 0 passed; 0 failed; 5 ignored; 0 measured; 0 filtered out; finished in 0.00s

     Running unittests (target/release/deps/tracing_macros-29cd8a3ad0edffa4)

running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.00s

     Running unittests (target/release/deps/trace-d249036429227e91)
WARNING: HTML report generation will become a non-default optional feature in Criterion.rs 0.4.0.
This feature is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'html_reports' feature in your Cargo.toml.

WARNING: In Criterion.rs 0.4.0, running criterion benchmarks outside of cargo-criterion will become a default optional feature.
The statistical analysis and reporting is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'cargo_bench_support' feature in your Cargo.toml.

Gnuplot not found, using plotters backend
otel_many_children/spec_baseline
                        time:   [22.107 us 22.173 us 22.262 us]
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  3 (3.00%) high severe
otel_many_children/no_data_baseline
                        time:   [27.640 us 27.707 us 27.771 us]
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild
otel_many_children/data_only_baseline
                        time:   [39.212 us 39.345 us 39.477 us]
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
otel_many_children/full time:   [100.89 us 101.07 us 101.29 us]
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe

     Running unittests (target/release/deps/tracing_serde-e0da2f22339ae83e)

running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.00s

     Running unittests (target/release/deps/enter-606f82b40ef7a1c2)
WARNING: HTML report generation will become a non-default optional feature in Criterion.rs 0.4.0.
This feature is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'html_reports' feature in your Cargo.toml.

WARNING: In Criterion.rs 0.4.0, running criterion benchmarks outside of cargo-criterion will become a default optional feature.
The statistical analysis and reporting is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'cargo_bench_support' feature in your Cargo.toml.

Gnuplot not found, using plotters backend
enter/enabled           time:   [13.157 ns 13.182 ns 13.208 ns]
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe
enter/disabled          time:   [1.8452 ns 1.8591 ns 1.8713 ns]
Found 8 outliers among 100 measurements (8.00%)
  8 (8.00%) low severe

enter_exit/enabled      time:   [68.054 ns 68.241 ns 68.459 ns]
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) high mild
  1 (1.00%) high severe
enter_exit/disabled     time:   [1.1779 ns 1.1822 ns 1.1888 ns]
Found 9 outliers among 100 measurements (9.00%)
  3 (3.00%) high mild
  6 (6.00%) high severe

enter_many/enabled      time:   [14.517 ns 14.538 ns 14.560 ns]
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe
enter_many/disabled     time:   [1.8624 ns 1.8761 ns 1.8886 ns]
Found 21 outliers among 100 measurements (21.00%)
  20 (20.00%) low severe
  1 (1.00%) low mild

     Running unittests (target/release/deps/filter-1aebdb22a20e3a3d)
WARNING: HTML report generation will become a non-default optional feature in Criterion.rs 0.4.0.
This feature is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'html_reports' feature in your Cargo.toml.

WARNING: In Criterion.rs 0.4.0, running criterion benchmarks outside of cargo-criterion will become a default optional feature.
The statistical analysis and reporting is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'cargo_bench_support' feature in your Cargo.toml.

Gnuplot not found, using plotters backend
static/baseline_single_threaded
                        time:   [81.353 ns 81.524 ns 81.707 ns]
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
static/single_threaded  time:   [40.788 ns 40.958 ns 41.126 ns]
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
static/single_threaded_reload
                        time:   [75.416 ns 75.689 ns 76.001 ns]
Found 23 outliers among 100 measurements (23.00%)
  18 (18.00%) low severe
  2 (2.00%) low mild
  2 (2.00%) high mild
  1 (1.00%) high severe
static/enabled_one      time:   [20.314 ns 20.339 ns 20.368 ns]
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low mild
  7 (7.00%) high mild
static/enabled_one_reload
                        time:   [37.948 ns 37.991 ns 38.041 ns]
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe
static/enabled_many     time:   [20.331 ns 20.371 ns 20.418 ns]
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
static/enabled_many_reload
                        time:   [37.965 ns 38.003 ns 38.043 ns]
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) low mild
static/disabled_level_one
                        time:   [233.66 ps 234.17 ps 234.71 ps]
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe
static/disabled_level_one_reload
                        time:   [467.37 ps 469.62 ps 472.37 ps]
Found 11 outliers among 100 measurements (11.00%)
  2 (2.00%) low mild
  4 (4.00%) high mild
  5 (5.00%) high severe
static/disabled_level_many
                        time:   [233.43 ps 234.03 ps 234.84 ps]
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe
static/disabled_level_many_reload
                        time:   [465.16 ps 466.08 ps 467.15 ps]
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
…