@hawkw 2022-05-06
> @hawkw should we close this?

I originally opened this PR because I was hoping to get some feedback from users and other maintainers about whether this kind of change would be desirable. I suppose we *could* interpret the lack of comments as a negative symbol, but I don't know...