@bryangarza 2022-05-07
Oh okay, gotcha, this feature does sound useful, but I can't really think of many concrete use cases 🤔 I can look at the code in more detail next week.