@HK416-is-all-you-need 2022-05-07
I will take a look at this, didn't notice you had PR idea, but I wonder if locking could be avoided compeltely